+++ src/main/java/co/bugg/quickplay/gui/PartyGui.java
@@ -47,7 +47,7 @@ public class PartyGui extends GuiScreen {
             drawDefaultBackground();
             mc.renderEngine.bindTexture(QuickPlay.icons.get(game.fileID));
             // Draw this games icon at the top
-            drawTexturedModalRect((width / 2 - Icons.iconWidth / 2), (float) (height * 0.05), game.xStart, game.yStart, Icons.iconWidth, Icons.iconHeight);
+            drawTexturedModalRect((width / 2 - Icons.iconWidth / 2), (height * 0.05f), game.xStart, game.yStart, Icons.iconWidth, Icons.iconHeight);
 
             // Draw the credits
             drawString(fontRendererObj, QuickPlay.credit, width / 2 - fontRendererObj.getStringWidth(QuickPlay.credit) / 2, height - 10, QuickPlay.configManager.getConfig().colors.get("primary").getRGB());
